backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and LangGraph dependencies
langgraph==0.2.16
langchain==0.2.16
langchain-core==0.2.38
langchain-aws==0.1.17

# AWS dependencies
boto3==1.35.36
botocore==1.35.36

# Data validation and parsing
pydantic==2.5.0
typing-extensions==4.8.0

# Environment management
python-dotenv==1.0.0

# Additional utilities
httpx==0.25.2
